backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
google-generativeai
python-dotenv
pytrends
pandas